bot/requirements.txt
CherryPy==18.8.0
emoji==2.2.0
numpy==1.23.5
python-dotenv==0.21.0
requests==2.28.1
telepot==12.7
